requirements.txt
pyperclip
fpdf2
jinja2